requirements.txt
streamlit==1.37.1
requests>=2.32.3
pandas>=2.2.2
plotly>=5.23.0
python-dotenv>=1.0.1
beautifulsoup4>=4.12.3
